Scrap current setup and prepare for modularisation
Add CommandWithSenderIter, which takes in IRC Messages as if by
Server::iter(), and yields a tuple of (Command, sender), as Command
doesn't have a sender associated with it.

Replace hardcoded if-based decision tree with a match-based one.
Extract decision tree into IrcResponder. Requires further modularisation.

Add remove command to help.

Take in non-optional slices in StarredMessage.